encapsulate the printing of results so it can be reused with any writer interface
